feat: add auth for endpoints

+++ src/main/java/com/example/myFirstProject/service/JournalEntryService.java
@@ -15,15 +15,16 @@ import java.util.Optional;
 // -> service code are the core business logic written in spring which is called by controller
 @Service
 public class JournalEntryService {
+
 JournalEntry journalEntry = new JournalEntry();
 @Autowired
     private JournalEntryRepository journalEntryRepository;
 //Autowiring means making a class from the bean which is already present in spring
     @Autowired
     private UserService userService;
+
     @Transactional
     public void saveEntry(JournalEntry journalEntry, String userName){
-        // holding it to user and calling the userservice to find the user by username
         User user = userService.findByUserName(userName);
         //set the localtime of the journal now so that API doesnt have to sent the date
         journalEntry.setDate(LocalDateTime.now());
@@ -32,22 +33,32 @@ JournalEntry journalEntry = new JournalEntry();
         // holding the saved journal entry so that user can be also added in that journal entry
         JournalEntry saved = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(saved);
-        userService.saveUser(user);
+        userService.saveNewUser(user);
         System.out.println("SAVED: " + saved);
     }
+
     public void saveEntry(JournalEntry journalEntry){
         journalEntryRepository.save(journalEntry);
     }
     public List<JournalEntry> getAll(){
         return journalEntryRepository.findAll();
     }
-    public Optional<JournalEntry> findById(ObjectId id){
+    public Optional<JournalEntry> findById(ObjectId id)
+    {
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        userService.saveUser(user);
-        journalEntryRepository.deleteById(id);
+
+    public void deleteById(ObjectId id, String userName) {
+        try {
+            User user = userService.findByUserName(userName);
+            boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An exception occured", e);
+        }
     }
 }
